refactors cli
