+++ conan/cli/commands/export_pkg.py
@@ -4,8 +4,9 @@ import os
 from conan.api.output import cli_out_write
 from conan.cli.command import conan_command, COMMAND_GROUPS
 from conan.cli.commands.install import _get_conanfile_path
-from conan.cli.common import get_lockfile, add_profiles_args, get_profiles_from_args, \
-    add_lockfile_args, add_reference_args, scope_options, save_lockfile_out
+from conan.cli.common import get_lockfile, get_profiles_from_args, \
+    scope_options, save_lockfile_out
+from conan.cli.args import add_lockfile_args, add_profiles_args, add_reference_args
 
 
 def json_export_pkg(info):
