+++ conan/cli/commands/source.py
@@ -2,7 +2,7 @@ import os
 
 from conan.cli.command import conan_command, COMMAND_GROUPS
 from conan.cli.commands.install import _get_conanfile_path
-from conan.cli.common import add_reference_args
+from conan.cli.args import add_reference_args
 from conan.api.conan_app import ConanApp
 from conans.client.conanfile.configure import run_configure_method
 from conans.client.graph.graph import CONTEXT_HOST
