+++ conan/cli/commands/test.py
@@ -4,9 +4,9 @@ from conan.api.output import ConanOutput
 from conan.cli.command import conan_command, COMMAND_GROUPS, OnceArgument
 from conan.cli.commands.create import test_package, _check_tested_reference_matches
 from conan.cli.commands.install import _get_conanfile_path
-from conan.cli.common import get_lockfile, get_profiles_from_args, _add_common_install_arguments, \
-    get_multiple_remotes, add_lockfile_args, save_lockfile_out
-from conan.cli.formatters.graph import print_graph_basic, print_graph_packages
+from conan.cli.common import get_lockfile, get_profiles_from_args, get_multiple_remotes, save_lockfile_out
+from conan.cli.args import add_lockfile_args, _add_common_install_arguments
+from conan.cli.printers.graph import print_graph_basic, print_graph_packages
 from conans.model.recipe_ref import RecipeReference
