+++ conans/test/unittests/client/command/parse_arguments_test.py
@@ -5,7 +5,7 @@ import unittest
 
 from parameterized.parameterized import parameterized_class
 
-from conan.cli.common import add_profiles_args
+from conan.cli.args import add_profiles_args
 
 
 @parameterized_class([{"argument": ["options", "-o", "--options"]},
